+++ CHANGELOG.md
@@ -5,6 +5,19 @@
 ### New features
 
 - Added `\Wizaplace\SDK\Basket\BasketItem::getProductCode`.
+- Updated `\Wizaplace\SDK\Catalog\Product` with `\Wizaplace\SDK\Division\Division` 
+- Added `\Wizaplace\SDK\Division\DivisionService::get`
+- Added `\Wizaplace\SDK\Company\CompanyService::getDivisionsCountriesCodes`
+- Added `\Wizaplace\SDK\Division\DivisionCompany`
+- Added `\Wizaplace\SDK\Company\CompanyService::getDivisions`
+- Added `\Wizaplace\SDK\Company\CompanyService::setDivisions`
+- Added `\Wizaplace\SDK\Pim\Product\ProductService::getDivisionsCountriesCodes`
+- Added `\Wizaplace\SDK\Pim\Product\ProductService::getDivisions`
+- Added `\Wizaplace\SDK\Pim\Product\ProductService::setDivisions`
+- Added `\Wizaplace\SDK\Division\DivisionService::set`
+- Added `\Wizaplace\SDK\User\UserAddress::getDivision`
+- Added `\Wizaplace\SDK\Catalog\ProductOffer`
+- Added `\Wizaplace\SDK\Catalog\Product::getOffers`
 
 ## 1.44.1